little housekeeping and version number bump

+++ KnodaIPhoneApp/KnodaIPhoneApp/Source/Categories/Core Categories/NSString+Utils.m
@@ -13,7 +13,7 @@
 - (NSString *)md5Hash {
     const char *cStr = [self UTF8String];
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    CC_MD5( cStr, strlen(cStr), result ); // This is the md5 call
+    CC_MD5( cStr, (unsigned int)strlen(cStr), result ); // This is the md5 call
     return [NSString stringWithFormat:
             @"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
             result[0], result[1], result[2], result[3],
